We have seen a speed-up after the initial switch https://github.com/dagger/dagger/pull/9887, and a slow-down after we went back https://github.com/dagger/dagger/pull/10086.

If we still see a speed-up after this switch, we are likely to want to stick with this for more than a few weeks.

As discussed with @sipsma in the last sync.

---

| Job Name | Before (P90 2w) | After (single run) |
|:--------|-------------------|-------------------|
| testdev-everything-else | 1016s (16m 56s) | `TODO` [FYI](https://github.com/dagger/dagger/pull/10248#issuecomment-2839991924) |
| testdev-module-runtimes | 851s (14m 11s) | 817s (13m 37s) |
| testdev-container | 768s (12m 48s) | 459s (7m 39s) |